Move the store into its own file
